LayoutTests/http/tests/websocket/tests/hybi/url-parsing-expected.txt

CONSOLE MESSAGE: Wrong url scheme for WebSocket http://127.0.0.1:8000/websocket/tests/hybi/null
CONSOLE MESSAGE: Invalid url for WebSocket ws://javascript:a
CONSOLE MESSAGE: Wrong url scheme for WebSocket http://127.0.0.1:8000/applet
CONSOLE MESSAGE: Wrong url scheme for WebSocket javascript:a
CONSOLE MESSAGE: URL has fragment component ws://127.0.0.1/path#
CONSOLE MESSAGE: URL has fragment component ws://127.0.0.1/path#fragment
Test WebSocket URL parsing.

On success, you will see a series of "PASS" messages, followed by "TEST COMPLETE".

PASS new WebSocket() threw exception TypeError: Not enough arguments.
PASS new WebSocket(null) threw exception SyntaxError: The string did not match the expected pattern..
PASS new WebSocket("ws://javascript:a") threw exception SyntaxError: The string did not match the expected pattern..
PASS new WebSocket("/applet") threw exception SyntaxError: The string did not match the expected pattern..
PASS new WebSocket("javascript:a") threw exception SyntaxError: The string did not match the expected pattern..
PASS (new WebSocket("ws://127.0.0.1:8880/a/../websocket/tests/hybi/simple")).url is "ws://127.0.0.1:8880/websocket/tests/hybi/simple"
PASS (new WebSocket("ws://127.0.0.1:8880/websocket/tests/hybi/simple?")).url is "ws://127.0.0.1:8880/websocket/tests/hybi/simple?"
PASS (new WebSocket("ws://127.0.0.1:8880/websocket/tests/hybi/simple?k=v")).url is "ws://127.0.0.1:8880/websocket/tests/hybi/simple?k=v"
PASS new WebSocket("ws://127.0.0.1/path#") threw exception SyntaxError: The string did not match the expected pattern..
PASS new WebSocket("ws://127.0.0.1/path#fragment") threw exception SyntaxError: The string did not match the expected pattern..
PASS successfullyParsed is true

TEST COMPLETE
